refactor(web): keep app router files route-thin
