+++ apps/web/src/app/(auth)/layout.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react';
-import { AuthShell } from '@/features/auth/presentation/auth-shell';
+import { AuthShell } from '@/modules/auth/presentation/auth-shell';
 
 const AuthLayout = ({ children }: { children: ReactNode }) => {
 	return <AuthShell>{children}</AuthShell>;
