+++ apps/web/src/app/(auth)/register/page.tsx
@@ -1,7 +1,3 @@
-import { RegisterForm } from '@/features/auth/presentation/register-form';
-
-const RegisterPage = () => {
-	return <RegisterForm />;
-};
+import { RegisterPage } from '@/modules/auth/presentation/register-page';
 
 export default RegisterPage;
