+++ apps/web/src/app/(dashboard)/client/orders/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { OrderDetailsPage } from '@/features/client-dashboard/presentation/order-details/order-details-page';
+import { OrderDetailsPage } from '@/modules/client-dashboard/presentation/order-details/order-details-page';
 
 type ClientOrderDetailsRouteProps = {
 	params: Promise<{
