+++ apps/web/src/app/(dashboard)/client/orders/new/page.tsx
@@ -1,4 +1,4 @@
-import { NewOrderPage } from '@/features/client-dashboard/presentation/new-order/new-order-page';
+import { NewOrderPage } from '@/modules/client-dashboard/presentation/new-order/new-order-page';
 
 const ClientNewOrderPage = () => {
 	return <NewOrderPage />;
