+++ apps/web/src/app/(dashboard)/client/page.tsx
@@ -1,4 +1,4 @@
-import { ClientDashboardPage } from '@/features/client-dashboard/presentation/overview/client-dashboard-page';
+import { ClientDashboardPage } from '@/modules/client-dashboard/presentation/overview/client-dashboard-page';
 
 const ClientPage = () => {
 	return <ClientDashboardPage />;
