+++ apps/web/src/app/(dashboard)/layout.tsx
@@ -1,6 +1,6 @@
 import { redirect } from 'next/navigation';
 import type { ReactNode } from 'react';
-import { DashboardShell } from '@/features/client-dashboard/presentation/dashboard/dashboard-shell';
+import { DashboardShell } from '@/modules/client-dashboard/presentation/dashboard/dashboard-shell';
 import { getAuthSession } from '@/shared/auth/session';
 
 const getClientSession = async () => {
