+++ apps/web/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { SmoothScroll } from '@packages/ui/providers/smooth-scroll';
-import { LandingPage } from '@/features/landing/presentation/landing-page';
+import { LandingPage } from '@/modules/landing/presentation/landing-page';
 
 export default function Home() {
 	return (
